flask_app/requirements.txt
Flask[async]
Pillow==10.4.0
schedule==1.2.2
ExifRead==3.0.0
APScheduler==3.10.4
waitress==3.0.0
